rag-workshop/requirements.txt
streamlit==1.32.0
langchain==0.1.5
langchain-google-genai==0.0.8
google-generativeai==0.3.2
chromadb==0.4.22
langchain-community==0.0.17 
pypdf==3.17.1
pandas==2.1.4
umap-learn==0.5.5
numpy==1.26.3
plotly==5.18.0
python-dotenv==1.0.0
python-multipart==0.0.20
